requirements/requirements.dev.txt
-r requirements.base.txt
coverage==4.5.4
flake8==3.7.7
model-mommy==1.6.0
pytest==4.4.1
pytest-cov==2.7.1
pytest-django==3.4.8
